chore: remove bump-engine step from RELEASING
