+++ dev/sdk.go
@@ -2,10 +2,14 @@ package main
 
 import (
 	"context"
+	"encoding/base64"
+	"fmt"
+	"strings"
 
 	"github.com/dagger/dagger/dev/internal/build"
 	"github.com/dagger/dagger/dev/internal/consts"
 	"github.com/dagger/dagger/dev/internal/dagger"
+	"github.com/moby/buildkit/identity"
 )
 
 // A dev environment for the official Dagger SDKs
@@ -91,4 +95,100 @@ func (dev *DaggerDev) introspection(ctx context.Context, installer func(*dagger.
 		WithFile("/usr/local/bin/codegen", builder.CodegenBinary()).
 		WithExec([]string{"codegen", "introspect", "-o", "/schema.json"}).
 		File("/schema.json"), nil
+}
+
+type gitPublishOpts struct {
+	sdk string
+
+	source, dest       string
+	sourceTag, destTag string
+	sourcePath         string
+	sourceFilter       string
+	sourceEnv          *dagger.Container
+
+	username    string
+	email       string
+	githubToken *dagger.Secret
+
+	dryRun bool
+}
+
+func gitPublish(ctx context.Context, opts gitPublishOpts) error {
+	base := opts.sourceEnv
+	if base == nil {
+		base = dag.Container().
+			From(consts.AlpineImage).
+			WithExec([]string{"apk", "add", "-U", "--no-cache", "git", "go", "python3"})
+	}
+
+	// FIXME: move this into std modules
+	git := base.
+		WithExec([]string{"git", "config", "--global", "user.name", opts.username}).
+		WithExec([]string{"git", "config", "--global", "user.email", opts.email})
+	if !opts.dryRun {
+		githubTokenRaw, err := opts.githubToken.Plaintext(ctx)
+		if err != nil {
+			return err
+		}
+		encodedPAT := base64.URLEncoding.EncodeToString([]byte("pat:" + githubTokenRaw))
+		git = git.
+			WithEnvVariable("GIT_CONFIG_COUNT", "1").
+			WithEnvVariable("GIT_CONFIG_KEY_0", "http.https://github.com/.extraheader").
+			WithSecretVariable("GIT_CONFIG_VALUE_0", dag.SetSecret("GITHUB_HEADER", fmt.Sprintf("AUTHORIZATION: Basic %s", encodedPAT)))
+	}
+
+	result := git.
+		WithEnvVariable("CACHEBUSTER", identity.NewID()).
+		WithWorkdir("/src/dagger").
+		WithExec([]string{"git", "clone", opts.source, "."}).
+		WithExec([]string{"git", "fetch", "origin", "-v", "--update-head-ok", fmt.Sprintf("refs/*%[1]s:refs/*%[1]s", strings.TrimPrefix(opts.sourceTag, "refs/"))}).
+		WithEnvVariable("FILTER_BRANCH_SQUELCH_WARNING", "1").
+		WithExec([]string{
+			"git", "filter-branch", "-f", "--prune-empty",
+			"--subdirectory-filter", opts.sourcePath,
+			"--tree-filter", opts.sourceFilter,
+			"--", opts.sourceTag,
+		})
+	if !opts.dryRun {
+		result = result.WithExec([]string{
+			"git",
+			"push",
+			// "--force", // NOTE: disabled to avoid accidentally rewriting the history
+			opts.dest,
+			fmt.Sprintf("%s:%s", opts.sourceTag, opts.destTag),
+		})
+	} else {
+		// on a dry run, just test that the last state of dest is in the current branch (and is a fast-forward)
+		history, err := result.
+			WithExec([]string{"git", "log", "--oneline", "--no-abbrev-commit"}).
+			Stdout(ctx)
+		if err != nil {
+			return err
+		}
+
+		destCommit, err := git.
+			WithEnvVariable("CACHEBUSTER", identity.NewID()).
+			WithWorkdir("/src/dagger").
+			WithExec([]string{"git", "clone", opts.dest, "."}).
+			WithExec([]string{"git", "switch", opts.destTag}).
+			WithExec([]string{"git", "rev-parse", "HEAD"}).
+			Stdout(ctx)
+		if err != nil {
+			if strings.Contains(err.Error(), "invalid reference: "+opts.destTag) {
+				// this is a ref that only exists in the source, and not in the
+				// dest, so no overwriting will occur
+				return nil
+			}
+			return err
+		}
+		destCommit = strings.TrimSpace(destCommit)
+
+		if !strings.Contains(history, destCommit) {
+			return fmt.Errorf("publish would rewrite history - %s not found\n%s", destCommit, history)
+		}
+		return nil
+	}
+
+	_, err := result.Sync(ctx)
+	return err
 }
